refacto App Skeleton to use SwiftUI Lifecycle
remove NSWindow, and old toolbar manager
