+++ RunningOrder/Shared/RunningOrderApp.swift
@@ -1,0 +1,55 @@
+//
+//  RunningOrderApp.swift
+//  RunningOrder
+//
+//  Created by Clément Nonn on 02/02/2021.
+//  Copyright © 2021 Worldline. All rights reserved.
+//
+
+import SwiftUI
+
+let changesService = CloudKitChangesService(container: CloudKitContainer.shared)
+
+@main
+struct RunningOrderApp: App {
+    @NSApplicationDelegateAdaptor private var appDelegate: AppDelegate
+
+    @StateObject var spaceManager = SpaceManager(
+        service: SpaceService(),
+        dataPublisher: changesService.spaceChangesPublisher.eraseToAnyPublisher()
+    )
+
+    @StateObject var sprintManager = SprintManager(
+        service: SprintService(),
+        dataPublisher: changesService.sprintChangesPublisher.eraseToAnyPublisher()
+    )
+
+    @StateObject var storyManager = StoryManager(
+        service: StoryService(),
+        dataPublisher: changesService.storyChangesPublisher.eraseToAnyPublisher()
+    )
+
+    @StateObject var storyInformationManager = StoryInformationManager(
+        service: StoryInformationService(),
+        dataPublisher: changesService.storyInformationChangesPublisher.eraseToAnyPublisher()
+    )
+
+    var body: some Scene {
+        WindowGroup {
+            MainView()
+                .environmentObject(spaceManager)
+                .environmentObject(sprintManager)
+                .environmentObject(storyManager)
+                .environmentObject(storyInformationManager)
+                .onAppear {
+                    appDelegate.changesService = changesService
+                    appDelegate.spaceManager = spaceManager
+                    changesService.fetchChanges()
+                }
+        }
+        .commands {
+            ToolbarCommands()
+            SidebarCommands()
+        }
+    }
+}
